move SqueezeExciteHparams
